+++ src/ngrest/base/NgRestModel.php
@@ -348,26 +348,7 @@ abstract class NgRestModel extends ActiveRecord implements GenericSearchInterfac
     }
     
     /**
-     * The NgRestFind is used when performing the crud list index overivew. You
-     * can override this method in order to hide data from the ngRestFind command
-     * which populates all data from the database.
-     *
-     * An example for hidding deleted news posts from the crud list:
-     *
-     * ```php
-     * public static function ngRestFind()
-     * {
-     *     return parent::ngRestFind()->where(['is_deleted' => 0]);
-     * }
-     * ```
-     *
-     * +
-     *
-     * > This method will taken for all internal *NOT API USER* related calls. So assuming an API user makes request to the APi
-     * > it will use find() instead of ngRestFind(). If a logged in user will make a request to an API it will take ngRestFind().
-     *
-     * @see {{yii\db\ActiveRecord::find()}}
-     * @return NgRestActiveQuery
+     * {@inheritDoc}
      */
     public static function ngRestFind()
     {